adds initialization for GOCELL-based model fitting
When GOCELL-based model fitting using the usual zeros-initialization fails, a retry is performed using an informed initialization based on morphological analysis of the image region. Since this also leads to a better initialization for GOCELLOS, which in turn then is more likely to succeed in optimization, it ultimately also avoids fallbacks to the GOCELL result.